Remove typegraphql-prisma in favor of graphql-code-generator. Also add MeiliSearch
