+++ .eslintignore
@@ -1,0 +1,3 @@
+src/generated
+node_modules
+dist
