+++ .prettierignore
@@ -1,3 +1,9 @@
 node_modules
 coverage
 dist
+src/generated
+.db-data
+.redis-data
+.meili-data
+cert
+.husky
