+++ Dockerfile
@@ -6,10 +6,16 @@ WORKDIR /usr/src/app
 
 COPY ./package.json .
 COPY ./package-lock.json .
-RUN npm install
+RUN npm ci
 
 COPY ./prisma .
-RUN npm run generate
+RUN npm run generate.prisma
+
+
+RUN mkdir src
+COPY ./codegen.yml .
+COPY ./src/schema.graphql ./src
+RUN npm run generate.graphql
 
 COPY . .
 RUN npm run build
