+++ codegen.yml
@@ -1,0 +1,11 @@
+overwrite: true
+schema: src/schema.graphql
+generates:
+  src/generated/graphql.ts:
+    plugins:
+      - "@graphql-codegen/typescript"
+      - "@graphql-codegen/typescript-resolvers"
+    config:
+      mappers:
+        User: .prisma/client#User as UserModel
+      inputMaybeValue: undefined | T
