+++ docker-compose.yml
@@ -43,6 +43,17 @@ services:
       - .redis-data:/data
     networks:
       - glimpse_network
+  search:
+    image: getmeili/meilisearch:v0.27.1
+    restart: unless-stopped
+    ports:
+      - "7700:7700"
+    env_file:
+      - .env
+    volumes:
+      - .meili-data:/meili_data
+    networks:
+      - glimpse_network
 networks:
   glimpse_network:
     name: glimpse_network
