+++ package.json
@@ -5,11 +5,13 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node dist/index.js",
-    "generate": "prisma generate",
+    "generate": "npm run generate.prisma && npm run generate.graphql",
+    "generate.prisma": "prisma generate",
+    "generate.graphql": "graphql-codegen",
     "build": "tsc",
     "docker.build": "docker build --tag rpitv/glimpse-api .",
-    "dev": "nodemon -L -e ts,js,json src/index.ts",
-    "dev.debug": "nodemon -L -e ts,js,json --exec \"node --inspect=0.0.0.0:5858 --require ts-node/register\" src/index.ts"
+    "dev": "concurrently \"nodemon -L -e ts,js,json src/index.ts\" \"npm run generate.graphql --watch\"",
+    "dev.debug": "concurrently \"nodemon -L -e ts,js,json --exec \\\"node --inspect=0.0.0.0:5858 --require ts-node/register\\\" src/index.ts\" \"npm run generate.graphql --watch\""
   },
   "repository": {
     "type": "git",
@@ -28,34 +30,35 @@
     "apollo-server-express": "^3.8.2",
     "argon2": "^0.28.5",
     "body-parser": "^1.19.0",
-    "class-validator": "^0.13.2",
     "connect-redis": "^6.1.3",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "fs-extra": "^8.1.0",
-    "graphql": "^15.8.0",
-    "graphql-fields": "^2.0.3",
+    "graphql": "^16.5.0",
+    "graphql-scalars": "^1.17.0",
     "graphql-type-json": "^0.3.2",
     "ip": "^1.1.8",
+    "meilisearch": "^0.25.1",
     "redis": "^4.1.0",
-    "reflect-metadata": "^0.1.13",
     "sharp": "^0.30.3",
-    "type-graphql": "^1.1.1",
     "typedi": "^0.10.0"
   },
   "devDependencies": {
+    "@graphql-codegen/cli": "2.6.2",
+    "@graphql-codegen/typescript": "2.5.0",
+    "@graphql-codegen/typescript-resolvers": "2.6.5",
     "@types/connect-redis": "^0.0.18",
     "@types/cors": "^2.8.12",
     "@types/express-session": "^1.17.4",
     "@types/fs-extra": "^8.1.1",
-    "@types/graphql-fields": "^1.3.4",
     "@types/ip": "^1.1.0",
     "@types/node": "^17.0.31",
     "@types/sharp": "^0.24.0",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
     "@typescript-eslint/parser": "^5.22.0",
+    "concurrently": "^7.2.1",
     "eslint": "^8.15.0",
     "eslint-config-prettier": "^8.5.0",
     "husky": "^8.0.0",
@@ -64,7 +67,6 @@
     "nodemon": "^2.0.7",
     "prettier": "^2.6.2",
     "prisma": "^3.14.0",
-    "ts-node": "^10.7.0",
-    "typegraphql-prisma": "^0.20.3"
+    "ts-node": "^10.7.0"
   }
 }
